update: user authentication for adding, editing and deleteing blogs

+++ routes/blogs.js
@@ -7,10 +7,11 @@ const { check } = require('express-validator');
 
 // Bring in models
 let Blog = require('../models/blog');
+let User = require('../models/user');
 
 
 // Add New Blog
-router.get('/add', (request, response) => {
+router.get('/add', ensureAuthenticated, (request, response) => {
     response.render('add_blog', {
         title: "Add Blog"
     });
@@ -19,7 +20,7 @@ router.get('/add', (request, response) => {
 // Add Submit POST Route
 router.post('/add',
     check('title').not().isEmpty().withMessage('Title is required'),
-    check('author').not().isEmpty().withMessage('Author is required'),
+    // check('author').not().isEmpty().withMessage('Author is required'),
     check('body').not().isEmpty().withMessage('Body is required'),
 
     // Get Error
@@ -34,7 +35,7 @@ router.post('/add',
         else {
             let blog = new Blog();
             blog.title = request.body.title;
-            blog.author = request.body.author;
+            blog.author = request.user.id;
             blog.body = request.body.body;
 
             blog.save((err) => {
@@ -51,8 +52,12 @@ router.post('/add',
 );
 
 // Load Edit Form
-router.get("/edit/:id", (request, response) => {
+router.get("/edit/:id", ensureAuthenticated, (request, response) => {
     Blog.findById(request.params.id, (err, blog) => {
+        if (blog.author != request.user.id) {
+            request.flash('danger', 'Not Authorised');
+            response.redirect('/');
+        }
         response.render('edit_blog', {
             title: 'Edit Title',
             blog: blog
@@ -82,24 +87,47 @@ router.post('/edit/:id', (request, response) => {
 
 // Delete Blog Route
 router.delete('/:id', (request, response) => {
+    if (!request.user._id) {
+        response.status(500).send();
+    }
+
     let query = { _id: request.params.id }
 
-    Blog.deleteOne(query, (err) => {
-        if (err) {
-            console.log(err);
+    Blog.findById(request.params.id, (err, blog) => {
+        if (blog.author != request.user._id) {
+            response.status(500).send();
+        } else {
+            Blog.deleteOne(query, (err) => {
+                if (err) {
+                    console.log(err);
+                }
+                request.flash('success', 'Blog Deleted');
+                response.send('Success');
+            });
         }
-        request.flash('success', 'Blog Deleted');
-        response.send('Success');
     });
 });
 
 // Get Single Blog
 router.get("/:id", (request, response) => {
     Blog.findById(request.params.id, (err, blog) => {
-        response.render('blog', {
-            blog: blog
+        User.findById(blog.author, (err, user) => {
+            response.render('blog', {
+                blog: blog,
+                author: user.name
+            });
         });
     });
 });
+
+// Access Control
+function ensureAuthenticated(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    } else {
+        req.flash('danger', 'Please Log In!');
+        res.redirect('/users/login');
+    }
+}
 
 module.exports = router;
